requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-multipart
